back-django/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==3.2.18
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
filelock==3.9.0
importlib-metadata==6.1.0
mysqlclient==2.1.0
pbr==5.11.1
Pillow==9.4.0
platformdirs==3.0.0
python-dateutil==2.8.2
pytz==2022.7.1
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
stevedore==3.5.2
typing_extensions==4.5.0
tzdata==2022.7
virtualenv==20.19.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
zipp==3.15.0
uWSGI